+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,7 @@
 
 Release History
 ===============
+* [BREAKING CHANGE] `role assignment delete`: prompt when delete all assignments under the subscription unconditionally
 * `ad app/sp credential reset`: support --end-date
 * `ad app permission add`: support to add permissions on existing api
 * `ad app permission list`: fix a bad error when there is no permissions